+++ src/main/java/com/fastcampus/investment/constant/UserInvestingType.java
@@ -1,4 +1,4 @@
-package com.fastcampus.investment.util.type;
+package com.fastcampus.investment.constant;
 
 public enum UserInvestingType {
     FAIL, CANCELED, INVESTED, NONE;
